newCommit
