+++ seleniumTests/README.md
@@ -1,0 +1,2 @@
+# Автотесты на java
+Для запуска надо [скачать последнюю версию chromedriver](https://chromedriver.chromium.org/downloads) и закинуть его в репозиторий.
