--- untitled/build.gradle
@@ -1,28 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'org.example'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    testImplementation 'org.seleniumhq.selenium:selenium-java:4.10.0'
-    testImplementation 'com.github.javafaker:javafaker:1.0.1'
-
-    compileOnly 'org.projectlombok:lombok:1.18.24'
-    annotationProcessor 'org.projectlombok:lombok:1.18.24'
-
-    testCompileOnly 'org.projectlombok:lombok:1.18.24'
-    testAnnotationProcessor 'org.projectlombok:lombok:1.18.24'
-}
-
-test {
-    useJUnitPlatform()
-    System.setProperty("webdriver.chrome.driver", "./driver/chromedriver.exe")
-}
